Refactor Reload Item, Change as business logic
- Use Item, Change Handlers to handle UI logic
- Refactor Reload Manager class to adhere to Single Responsibility Principle and have separate classes for Item Reloader & Change Reloader to handle the related business logic
- Remove classes no longer used
- Add missing spec to Vending Machine & modify relate existing spec to use Item Reloader
- Maker other related changes where applicable
