+++ lib/cli/item_reload_handler.rb
@@ -1,4 +1,4 @@
-class ItemReloader
+class ItemReloadHandler
   def initialize(vending_machine, display, input_handler)
     @vending_machine = vending_machine
     @display = display
